fixed photon samples in defineDatasets

+++ ConfigParser.C
@@ -10,7 +10,7 @@
 using namespace std;
 
 bool isWhiteSpace(string str){
-	for (int i = 0; i<str.size(); i++){
+	for (int i = 0; i<(int)str.size(); i++){
 		if ( (! (str[i] == ' ')) && (! (str[i] == '\t') ) ){
 			return false;	
 		}
